requirements.txt
Django==1.8.4
django-extensions==1.6.1
django-filter==0.13.0
djangorestframework==3.3.2
ecdsa==0.13
Fabric==1.12.0
paramiko==1.17.2
pycrypto==2.6.1
redis==2.10.5
six==1.10.0
wheel==0.24.0
